restructure order aggregate
