+++ order/build.gradle
@@ -1,0 +1,3 @@
+bootJar {
+    enabled = false
+}
