+++ order/query-side/src/main/java/org/example/event/sourcing/order/poc/query/order/domain/entity/OrderRecord.java
@@ -1,0 +1,34 @@
+package org.example.event.sourcing.order.poc.query.order.domain.entity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import java.time.Instant;
+
+@Entity
+@Table(name = "ORDER_RECORD")
+@EntityListeners(AuditingEntityListener.class)
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class OrderRecord {
+
+    @Id
+    private String orderId;
+
+    private OrderStatus status;
+
+    private Instant createdDate;
+
+    private Instant updatedDate;
+
+}
