+++ order/query-side/src/main/java/org/example/event/sourcing/order/poc/query/order/domain/entity/OrderStatus.java
