+++ settings.gradle
@@ -1,13 +1,17 @@
 rootProject.name = 'event-sourcing-order-poc'
 
 include 'common'
-include 'command-side-order'
-include 'event-handler-order'
-include 'query-side-order'
 include 'command-side-payment'
 include 'event-handler-payment'
 include 'query-side-payment'
 include 'command-side-shipment'
 include 'event-handler-shipment'
 include 'query-side-shipment'
+include 'order'
+include 'order:command-side'
+findProject(':order:command-side')?.name = 'command-side'
+include 'order:event-handler'
+findProject(':order:event-handler')?.name = 'event-handler'
+include 'order:query-side'
+findProject(':order:query-side')?.name = 'query-side'
